+++ Network/include/IConnectionController.h
@@ -141,7 +141,7 @@ public:
 	 * @param p_Objects array of objects to be removed
 	 * @param p_NumObjects the number of objects in the array
 	 */
-	virtual void sendRemoveObjects(const uint16_t* p_Objects, unsigned int p_NumObjects) = 0;
+	virtual void sendRemoveObjects(const uint32_t* p_Objects, unsigned int p_NumObjects) = 0;
 
 	/**
 	 * Get the number of objects in the package.
@@ -157,7 +157,7 @@ public:
 	 * @param p_Package a valid reference to a package with the RemoveObjects type.
 	 * @return an array of object references to objects to remove
 	 */
-	virtual const uint16_t* getRemoveObjectRefs(Package p_Package) = 0;
+	virtual const uint32_t* getRemoveObjectRefs(Package p_Package) = 0;
 
 	/**
 	 * Send an Object Action package.
